backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
python-dotenv==1.0.1
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
chromadb==0.5.5
sentence-transformers==3.1.0
PyMuPDF==1.24.10
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
aiofiles==24.1.0
